+++ src/main/kotlin/app/hinl/oreharvest/screenhandler/ButtonEntry.kt
@@ -1,0 +1,5 @@
+package app.hinl.oreharvest.screenhandler
+
+import net.minecraft.text.Text
+
+data class ButtonEntry(val isChecked: Boolean, val buttonText: Text, val buttonAction: (Boolean) -> Unit)