+++ src/main/kotlin/app/hinl/oreharvest/utils/SyncPair.kt
@@ -1,0 +1,3 @@
+package app.hinl.oreharvest.utils
+
+data class SyncPair(val setFunc: (Any) -> Unit, val getFunc: (Any) -> Unit)